@@ -5,7 +5,7 @@ publish_to: 'none'
 version: 1.0.0+1
 
 environment:
-  sdk: ">=2.7.0 <3.0.0"
+  sdk: ">=2.12.0 <3.0.0"
 
 dependencies:
   flutter:
@@ -13,7 +13,7 @@ dependencies:
   
   story_painter:
     path: ../
-  fast_color_picker: 0.0.1
+  fast_color_picker: 0.1.1
 cupertino_icons: ^0.1.3
 
 dev_dependencies:
