@@ -1,16 +1,16 @@
 name: story_painter
 description: This package provides drawing experince like in story editor of Instagram.
-version: 0.0.2
+version: 0.0.3
 homepage: https://github.com/furkansarihan/story_painter
 
 environment:
-  sdk: ">=2.7.0 <3.0.0"
-  flutter: ">=1.17.0 <2.0.0"
+  sdk: '>=2.12.0 <3.0.0'
+  flutter: ">=2.0.0"
 
 dependencies:
   flutter:
     sdk: flutter
-  flutter_svg: ^0.19.0
+  flutter_svg: ^0.22.0
 
 dev_dependencies:
   flutter_test:
