@@ -111,17 +111,17 @@ class PathUtil {
   }
 
   static List<T> translate<T extends Offset>(List<T> data, Offset location) {
-    final output = List<T>();
+    final output = <T>[];
 
     data.forEach(
-        (point) => output.add(point.translate(location.dx, location.dy)));
+        (point) => output.add(point.translate(location.dx, location.dy) as T));
 
     return output;
   }
 
   static List<List<T>> translateData<T extends Offset>(
       List<List<T>> data, Offset location) {
-    final output = List<List<T>>();
+    final output = <List<T>>[];
 
     data.forEach((set) => output.add(translate(set, location)));
 
@@ -129,16 +129,16 @@ class PathUtil {
   }
 
   static List<T> scale<T extends Offset>(List<T> data, double ratio) {
-    final output = List<T>();
+    final output = <T>[];
 
-    data.forEach((point) => output.add(point.scale(ratio, ratio)));
+    data.forEach((point) => output.add(point.scale(ratio, ratio) as T));
 
     return output;
   }
 
   static List<List<T>> scaleData<T extends Offset>(
       List<List<T>> data, double ratio) {
-    final output = List<List<T>>();
+    final output = <List<T>>[];
 
     data.forEach((set) => output.add(scale(set, ratio)));
 
@@ -146,7 +146,7 @@ class PathUtil {
   }
 
   static List<T> normalize<T extends Offset>(List<T> data,
-      {Rect bound, double border}) {
+      {Rect? bound, double? border}) {
     bound ??= bounds(data);
     border ??= 0.0;
 
@@ -157,7 +157,7 @@ class PathUtil {
   }
 
   static List<List<T>> normalizeData<T extends Offset>(List<List<T>> data,
-      {Rect bound}) {
+      {Rect? bound}) {
     bound ??= boundsOf(data);
 
     final ratio = 1.0 / max(bound.width, bound.height);
@@ -169,7 +169,7 @@ class PathUtil {
   }
 
   static List<T> fill<T extends Offset>(List<T> data, Rect rect,
-      {Rect bound, double border}) {
+      {Rect? bound, double? border}) {
     bound ??= bounds(data);
     border ??= 4.0;
 
@@ -202,7 +202,7 @@ class PathUtil {
   }
 
   static List<List<T>> fillData<T extends Offset>(List<List<T>> data, Rect rect,
-      {Rect bound, double border}) {
+      {Rect? bound, double? border}) {
     bound ??= boundsOf(data);
     border ??= 4.0;
 
@@ -246,7 +246,7 @@ class PathUtil {
   }
 
   static List<Path> toPaths(List<List<Offset>> data) {
-    final paths = List<Path>();
+    final paths = <Path>[];
 
     data.forEach((line) => paths.add(toPath(line)));
 
@@ -281,7 +281,7 @@ class PathUtil {
   }
 
   static List<Path> scalePaths(List<Path> data, double ratio) {
-    final output = List<Path>();
+    final output = <Path>[];
 
     data.forEach((path) => output.add(scalePath(path, ratio)));
 
@@ -289,7 +289,7 @@ class PathUtil {
   }
 
   static List<Path> translatePaths(List<Path> data, Offset location) {
-    final output = List<Path>();
+    final output = <Path>[];
 
     final transform = Matrix4.identity();
     transform.translate(location.dx, location.dy);
@@ -300,7 +300,7 @@ class PathUtil {
   }
 
   static List<Path> parseDrawable(DrawableParent root) {
-    final list = List<Path>();
+    final list = <Path>[];
 
     _parseDrawableRoot(root, list);
 
@@ -309,7 +309,7 @@ class PathUtil {
 
   static _parseDrawableRoot(DrawableParent root, List<Path> output) {
     if (root.children != null) {
-      root.children.forEach((drawable) {
+      root.children!.forEach((drawable) {
         if (drawable is DrawableShape) {
           output.add(drawable.path);
         } else if (drawable is DrawableParent) {
@@ -321,7 +321,7 @@ class PathUtil {
 
   static Size getDrawableSize(DrawableRoot root) => root.viewport.size;
 
-  static Path toShapePath(List<CubicLine> lines, double size, double maxSize) {
+  static Path toShapePath(List<CubicLine> lines, double size, double? maxSize) {
     assert(lines.length > 0);
 
     if (lines.length == 1) {
@@ -333,13 +333,13 @@ class PathUtil {
           ..line(line.end);
       }
 
-      return line.toShape(size, maxSize);
+      return line.toShape(size, maxSize!);
     }
 
     final path = Path();
 
     final firstLine = lines.first;
-    path.start(firstLine.start + firstLine.cpsUp(size, maxSize));
+    path.start(firstLine.start + firstLine.cpsUp(size, maxSize!));
 
     for (int i = 0; i < lines.length; i++) {
       final line = lines[i];
